Show DragonStyle in UI
Crane Style showed an icon in the UI, but Dragon Style did not.
It is possible to quickly toggle Crane Style and Dragon Style out
of combat to have them both show as on, even though actually only
one is active. Showing a buff allows the player to see which buff
is actually active if both seem like they are on.

Signed-off-by: Benjamin Fennema <benjamin.fennema@gmail.com>